==> notebooks/00_exploration.ipynb <==
DATA_PATH = "../data/raw/survey_results_public_2022.csv"
EXPORT_PATH = "../data/processed/0_preprocessed_df.pkl"

import pandas as pd 
import numpy as np
import pickle

pd.options.display.max_rows = 100
pd.options.display.max_columns = 100

# Read data and print shape
raw_df = pd.read_csv(DATA_PATH)
raw_df.shape

#Display a snippet of the raw data
raw_df.head(3)

# Display random answer 
raw_df.sample(1).iloc[0]

# Print the general information of the data frame 
raw_df.info()

# Get stats for the numerical column
raw_df.describe()

# Investigate the questionable objects columns
questionable_cols = ["YearsCodePro", "YearsCode", "EdLevel"]

for col in questionable_cols: 
    print(f"No of unique values for '{col}' are {raw_df[col].nunique()}")
    print(raw_df[col].unique().tolist())
    print('--------------------------')
    print()

#Columns to keep for further work
cols_to_keep = ["Employment",
               "RemoteWork",
               "EdLevel",
               "YearsCode",
               "YearsCodePro",
               "DevType",
               "Country",
               "CompTotal",
               "LanguageHaveWorkedWith",
               "LanguageWantToWorkWith",
               "DatabaseHaveWorkedWith",
               "DatabaseWantToWorkWith",
               "PlatformHaveWorkedWith",
               "PlatformWantToWorkWith",
               "WebframeHaveWorkedWith",
               "WebframeWantToWorkWith",
               "MiscTechHaveWorkedWith",
               "MiscTechWantToWorkWith",
               "ToolsTechHaveWorkedWith",
               "ToolsTechWantToWorkWith",
               "NEWCollabToolsHaveWorkedWith",
               "NEWCollabToolsWantToWorkWith",
               "OpSysProfessional use",
               "OpSysPersonal use",
               "VersionControlSystem",
               "VCInteraction",
               "Gender",
               "WorkExp",
               ]

df = raw_df[cols_to_keep]
df.head()

#Export Data
df.to_pickle(EXPORT_PATH)
