==> notebooks/01_preprocessing.ipynb <==
# Constants
# DATA_PATH   = "../data/raw/survey_results_public_2022.csv"
DATA_PATH   = "../data/processed/0_preprocessed_df.pkl"
EXPORT_PATH = "../data/processed/01_preprocessed_df.pkl"

REPLACE_DICT = {
    'YearsCodePro': {'Less than 1 year': 0, 'More than 50 years': 51}, 
    'YearsCode':    {'Less than 1 year': 0, 'More than 50 years': 51}}


# Load packages
import pandas as pd 
import numpy as np
import logging
import pickle

___________

## Functions

def split_answers(data_series, delimiter=";"):
    """ 
    Split multiple answers in a single string 
    to a list of single strings each represnting a single answers 

    Parameters:
    * data_series (pd.Series): String series with answers 
    * delimiter (string): Another decimal integer 
                          Defaults to ";"

    Returns: (pd.Series) 
    """

    # Check if multiple answers exist - if none: return original 
    if not data_series.str.contains(delimiter).any():
        return data_series
    
    # Else, split each value to a list 
    modified_series = data_series.str.split(delimiter)    
    
    # Replace NAs with empty lists 
    null_mask = modified_series.isnull()
    modified_series.loc[null_mask] = modified_series.loc[null_mask].apply(lambda x: [])
    
    return modified_series

___________

## Processing 

#processing data
df = pd.read_pickle(DATA_PATH)
df1 = df.copy()

df1.head(3)

#Replace Values with the dictionary
for col, replacement in REPLACE_DICT.items():
    df1[col] = df1[col].replace(replacement)

#Split multiple answers
cat_cols = df1.select_dtypes(include = "object")
for col in cat_cols:
    df1[col] = split_answers(df1[col])

df1.head()

___________

## Visually verify results 

i = df.sample(1).index[0]
print(df['LanguageHaveWorkedWith'].iloc[i])
print(df1['LanguageHaveWorkedWith'].iloc[i])

i = df.sample(1).index[0]
print(df['DevType'].iloc[i])
print(df1['DevType'].iloc[i])

i = df.sample(1).index[0]
print(df['EdLevel'].iloc[i])
print(df1['EdLevel'].iloc[i])

i = df.sample(1).index[0]
print(df['YearsCodePro'].iloc[i])
print(df1['YearsCodePro'].iloc[i])

## Export Data 

df1.to_pickle(EXPORT_PATH)
