==> notebooks/04b_model_job_baseline.ipynb <==
DF_PATH       = "../data/processed/2_cleaned_data.pkl"

MLFLOW_TRACKING_URI = '../models/mlruns'
MLFLOW_EXPERIMENT_NAME = "skills_jobs_stackoverflow"

LOG_PATH = "../models/temp/"
LOG_DATA_PKL    =  "data.pkl"
LOG_MODEL_PKL   =  "model.pkl"
LOG_METRICS_PKL =  "metrics.pkl"

# Load packages
import pandas as pd 
import numpy as np
import logging
import pickle
import random
# import plotly 
import os
from pathlib import Path

import mlflow
from mlflow.tracking import MlflowClient

from sklearn.preprocessing import StandardScaler
from sklearn.pipeline import make_pipeline
from sklearn.multioutput import MultiOutputClassifier
from sklearn.model_selection import train_test_split

from sklearn import metrics
from sklearn.metrics import auc, accuracy_score, confusion_matrix, f1_score, precision_score, recall_score

from sklearn.linear_model import LogisticRegression

_________

### Functions

def calculate_performance(ground_truth, predictions, metric_function, sort_values=False):
    performance_scores = {}
    #loop over roles
    for col in predictions.columns:
        role_pred  = predictions[col].copy()
        role_truth = ground_truth[col].copy()
        #calculate metric performance for the role
        performance_scores[col] = round(metric_function(role_truth, role_pred) * 100, 2)
        
    performance_scores = pd.Series(performance_scores.values(), index=performance_scores.keys())
    if sort_values:
        performance_scores = performance_scores.sort_values()
    
    return performance_scores

_________

# Initialize

### Create directories

Path(MLFLOW_TRACKING_URI).mkdir(parents=True, exist_ok=True)
Path(LOG_PATH).mkdir(parents=True, exist_ok=True)

### Read data

# Read Data
df = pd.read_pickle(DF_PATH)
df

### Balance classes 

# Check the total samples of roles
roles_df = df["DevType"].copy()
roles_df.sum(axis=0)

print(f"Shape of roles dataframe : {roles_df.shape}")
roles_df.head(3)

#Resample roles
samples_per_class = 1200
resampled_roles = []

for role in roles_df.columns:
    #get the role samples
    sub_df = roles_df.loc[roles_df[role]==1].copy()
    
    if len(sub_df) < samples_per_class:
        #Upsample
        sub_df = sub_df.sample(samples_per_class, replace=True, random_state=11)
    else:
        #Downsample
        sub_df = sub_df.sample(samples_per_class, random_state=11)
        
    resampled_roles.append(sub_df)

resampled_roles_df = pd.concat(resampled_roles)
print(f"Shape of resampled roles dataframe : {resampled_roles_df.shape}")
resampled_roles_df.head(3)

resampled_roles_df.sum(axis=0)

# Construct dfs
# roles_df  = pd.concat(resampled_roles)
df = df.loc[roles_df.index].copy()
df.head(3)

### Split

X_train, X_test, Y_train, Y_test = train_test_split(df.drop("DevType", axis=1), 
                                                    df["DevType"], 
                                                    random_state=11)

____________

# Train models

### Initialize MLflow

# Initialize client and experiment
mlflow.set_tracking_uri(MLFLOW_TRACKING_URI)
mlflow.set_experiment(MLFLOW_EXPERIMENT_NAME)
client = MlflowClient()
mlflow.get_experiment_by_name(MLFLOW_EXPERIMENT_NAME)
exp = client.get_experiment_by_name(MLFLOW_EXPERIMENT_NAME)

### 1. Logistic regression

clf = make_pipeline(StandardScaler(),
                    MultiOutputClassifier(LogisticRegression()))

clf.fit(X_train, Y_train)

# Evaluate on training set
predictions =  pd.DataFrame(clf.predict(X_train),
                            columns=Y_train.columns)
train_scores = {score.__name__: calculate_performance(Y_train, predictions, score) 
                for score in [accuracy_score, precision_score, recall_score, f1_score]}
train_scores = pd.concat(train_scores,axis=1)

# Evaluate on test set
predictions =  pd.DataFrame(clf.predict(X_test), columns=Y_test.columns)
test_scores = {score.__name__: calculate_performance(Y_test, predictions, score) 
                for score in [accuracy_score, precision_score, recall_score, f1_score]}
test_scores = pd.concat(test_scores,axis=1)
mean_test_scores = test_scores.mean()

print(mean_test_scores)
test_scores

## Log run

### 1. Prepare

# Data details
data_details = {"data_path": DF_PATH,
                "training_indices": X_train.index.tolist(),
                "test_indices":     X_test.index.tolist(), 
                "features_names":   X_train.columns.droplevel(0).tolist(),
                "targets_names":    Y_train.columns.tolist()}

with open(os.path.join(LOG_PATH, LOG_DATA_PKL), "wb") as output_file:
    pickle.dump(data_details, output_file)

# Model
model = {"model_description": "Baseline model: Logistic Regression ",
         "model_details": str(clf),
         "model_object": clf} 

with open(os.path.join(LOG_PATH, LOG_MODEL_PKL), "wb") as output_file:
    pickle.dump(model, output_file)

# Performance details
classes_metrics = {"train_scores": train_scores, 
                   "test_scores":  test_scores}

with open(os.path.join(LOG_PATH, LOG_METRICS_PKL), "wb") as output_file:
    pickle.dump(classes_metrics, output_file)

### 2. Log

# Start a new run and track 
with mlflow.start_run(experiment_id=exp.experiment_id, run_name=model["model_description"]):
    # Log pickles 
    mlflow.log_artifacts(LOG_PATH)
    
    # Track metrics 
    for metric, score in mean_test_scores.items():
        mlflow.log_metric(metric, score)
