==> notebooks/06_job_prediction.ipynb <==
MLFLOW_TRACKING_URI = '../models/mlruns/'
MLFLOW_RUN_ID = "31ae402b0ee14d82934009728f2c3caf"

LOG_DATA_PKL    =  "data.pkl"
LOG_MODEL_PKL   =  "model.pkl"
LOG_METRICS_PKL =  "metrics.pkl"

CLUSTERS_YAML_PATH = "../data/processed/features_skills_clusters_description.yaml"

import os 
import sklearn
import pickle
import yaml

import pandas as pd

import mlflow
from mlflow.tracking import MlflowClient

# Initialize

# 1. Mlflow

# Initialize client and experiment
mlflow.set_tracking_uri(MLFLOW_TRACKING_URI)
client = MlflowClient()

run = mlflow.get_run(MLFLOW_RUN_ID)
artificats_path = run.info.artifact_uri

## Load model

# Load model
model_path = os.path.join(artificats_path, LOG_MODEL_PKL).replace("file:///", "")
with open(model_path, "rb") as f:
    model = pickle.load(f)

model

# Load data pkl
data_path  = os.path.join(artificats_path, LOG_DATA_PKL).replace("file:///", "")
with open(data_path, 'rb') as data_file:
    data = pickle.load(data_file)

data.keys()

# Unpack vars
features_names = pd.Series(data["features_names"])
targets_names  = pd.Series(data['targets_names'])
classifier = model['model_object']

classifier

______

## Load skills Clusters

# Load skills clusters
with open(CLUSTERS_YAML_PATH, "r") as clusters_file:
    skills_clusters = yaml.safe_load(clusters_file)

skills_clusters

# Reformat into data frame
molten_clusters = [(cluster_name, cluster_skill)
                   for cluster_name, cluster_skills in skills_clusters.items()
                   for cluster_skill in cluster_skills]

clusters_df = pd.DataFrame(molten_clusters, columns=["cluster_name", "skill"])
clusters_df

________

## Predict sample entry

sample_skills = ['NumPy', 'Pandas', 'Scikit-learn']

# Verify
pd.Series(sample_skills).isin(features_names)

### 1. Recreate cluster features

sample_clusters = clusters_df.copy()
sample_clusters["sample_skills"] = sample_clusters["skill"].isin(sample_skills)
sample_clusters

cluster_features = sample_clusters.groupby("cluster_name")["sample_skills"].sum()
cluster_features

### 2. Create OneHotEncoded skills

#Exclude clusters names & get only skills names
skills_names = features_names[~features_names.isin(cluster_features.index)]
skills_names

ohe_skills = pd.Series(skills_names.isin(sample_skills).astype(int).tolist(), 
                       index=skills_names)
ohe_skills

### 3. Combine features

# Concat
features = pd.concat([ohe_skills,
                      cluster_features])
features

# Sort columns
features = features.loc[features_names]
features

### 4. Predict

prob_predictions = classifier.predict_proba([features.values])
prob_predictions

positive_probs = [prob[0][1] for prob in prob_predictions]
job_prediction = pd.Series(positive_probs, 
                           index=targets_names).sort_values(ascending=False)

job_prediction
